upsolve: board: marginal speed improvement (~5%)

+++ common/src/lib.rs
@@ -2,7 +2,7 @@
 pub mod api;
 pub mod board;
 pub mod evaluate;
-pub mod f64;
+pub mod float;
 pub mod geom;
 pub mod problem;
 pub mod board_options;
